moved menu button into its own component
