+++ components/HeaderNoBackButton.tsx
@@ -1,21 +1,13 @@
 import React from 'react';
-import { Pressable, View } from "react-native";
-import { Ionicons } from '@expo/vector-icons';
+import { View } from "react-native";
 import styles from '../styles/header';
+import MenuIcon from './MenuIcon';
 
 export default function HeaderNoBackButton() {
-
-  const handleToggleMenu = () => {
-    console.log('menu toggled');
-  }
   
   return (
     <View style={{...styles.header, ...styles.headerNoBackButton}}>
-      <Pressable onPressOut={handleToggleMenu}>
-        <Ionicons 
-          name="ios-menu-outline"
-          size={30} />
-      </Pressable>
+      <MenuIcon />
     </View>
   )
 }
