+++ components/HeaderWithBackButton.tsx
@@ -1,23 +1,16 @@
 import React from 'react';
-import { Pressable, View } from "react-native";
-import { Ionicons } from '@expo/vector-icons';
+import { View } from "react-native";
 import BackButton from './BackButton';
 import styles from '../styles/header';
+import MenuIcon from './MenuIcon';
 
 export default function HeaderWithBackButton() {
 
-  const handleToggleMenu = () => {
-    console.log('menu toggled');
-  }
   
   return (
     <View style={{...styles.header, ...styles.headerWithBackButton}}>
       <BackButton />
-      <Pressable onPressOut={handleToggleMenu}>
-        <Ionicons 
-          name="ios-menu-outline"
-          size={30} />
-      </Pressable>
+      <MenuIcon />
     </View>
   )
 }
